requirements.txt
line-bot-sdk==3.7
flask==3.0.0
gunicorn
python-dotenv
psycopg2-binary
requests